parser/requirements.txt
beautifulsoup4~=4.11.1
lxml~=4.9.1
requests~=2.28.1
selenium~=4.4.3
webdriver-manager~=3.8.3
mysql-connector-python~=8.0.30